+++ virtual-keyboard/index.html
@@ -4,8 +4,9 @@
         <meta charset="utf-8">
         <meta name="viewport" content="width=device-width,initial-scale=1">
         <title>Keyboard</title>
-    <script type="text/javascript" src="js/index.js"></script>
-    <link href="css/index.css" rel="stylesheet">
+        <script type="module" src="js/key.js"></script>
+        <script type="module" src="js/index.js"></script>
+        <link href="css/index.css" rel="stylesheet">
 </head>
     <body>
     </body>
