+++ virtual-keyboard/js/index.js
@@ -1,26 +1,4 @@
-const KEY_OF_KEYBOARD = {
-    EN_LOW_WORD_ARRAY : ['§','1','2','3','4','5','6','7','8','9','0','-','=',
-                        '&#5130;','&#8644;','q','w','e','r','t','y','u','i','o','p','[',']',"EN",
-                        '&#8682;','a','s','d','f','g','h','j','k','l',';','\'','\\','&#8629;',
-                        '&#8679;','`','z','x','c','v','b','n','m',',','.','/','↑','&#8679;',
-                        'control','option','command','MacOS','command','option','←','↓','→'],
-    EN_UP_WORD_ARRAY : ['±','!','@','#','$','%','^','&','*','(',')','-','=','&#5130;',
-                        '&#8644;','Q','W','E','R','T','Y','U','I','O','P','[',']',"EN",
-                        '&#8682;','A','S','D','F','G','H','J','K','L',';','\'','\\','&#8629;',
-                        '&#8679;','`','Z','X','C','V','B','N','M ',',','.','/','↑','&#8679;',
-                        'control','option','command','MacOS','command','option','←','↓','→'],
-    LG_LOW_WORD_ARRAY : ['§','1','2','3','4','5','6','7','8','9','0','-','=','&#5130;',
-                        '&#8644;','й','ц','у','к','е','н','г','i','ў','з','х','ъ',"BY",
-                        '&#8682;','ф','ы','в','а','п','р','о','л','д','ж','э','ё','&#8629;',
-                        '&#8679;','`','я','ч','с','м','і','т','ь','б','ю','/','↑','&#8679;',
-                        'control','option','command','MacOS','command','option','←','↓','→'],
-    LG_UP_WORD_ARRAY : ['>','!','"','№','%',':',',','.',';','(',')','-','=','&#5130;',
-                        '&#8644;','Й','Ц','У','К','Е','Н','Г','I','Ў','З','Х','Ъ',"BY",'&#8682;',
-                        'Ф','Ы','В','А','П','Р','О','Л','Д','Ж','Э','Ё','&#8629;',
-                        '&#8679;','`','Я','Ч','С','М','І','Т','Ь','Б','Ю','/','↑','&#8679;',
-                        'control','option','command','MacOS','command','option','←','↓','→']
-
-                    };
+import {KEY_OF_KEYBOARD} from './js/key.js'
 
 class Key {
     constructor(element, elClass, elName, value, children){
